+++ rust_library/keychainstore/build.gradle.kts
@@ -6,7 +6,6 @@ import gobley.gradle.Variant
 import gobley.gradle.cargo.dsl.jvm
 import gobley.gradle.rust.targets.RustPosixTarget
 import gobley.gradle.rust.targets.RustWindowsTarget
-import gobley.gradle.uniffi.tasks.BuildBindingsTask
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 
 plugins {
@@ -22,10 +21,6 @@ plugins.withId("publish-plugin") {
   project.description = "desktop/ios平台密钥存储模块"
   project.version = "1.2.0"
 }
-
-val isPublish =
-  gradle.startParameter.taskNames.any { it.endsWith("publish") || it.endsWith("publishToMavenLocal") }
-
 
 kotlin {
   jvm("desktop")
@@ -111,7 +106,9 @@ uniffi {
 
 tasks.named("compileKotlinDesktop") {
   doFirst {
-    projectDir.resolve("src").deleteRecursively()
+    if (!project.isPublish) {
+      projectDir.resolve("src").deleteRecursively()
+    }
   }
   doLast {
     copyDirectoryToTarget(
@@ -155,7 +152,7 @@ tasks.named("gen-bindings") {
 project.afterEvaluate {
   tasks.named("buildBindings") {
     doLast {
-      if (isPublish) {
+      if (project.isPublish) {
         projectDir.resolve("build").resolve("generated").resolve("uniffi").listFiles().forEach {
           if (it.path.contains("Main")) {
             it.deleteRecursively()
