+++ rust_library/ziplib/build.gradle.kts
@@ -6,7 +6,6 @@ import gobley.gradle.Variant
 import gobley.gradle.cargo.dsl.jvm
 import gobley.gradle.rust.targets.RustAndroidTarget
 import gobley.gradle.rust.targets.RustPosixTarget
-import gobley.gradle.uniffi.tasks.BuildBindingsTask
 import org.gradle.kotlin.dsl.support.serviceOf
 import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
@@ -26,9 +25,6 @@ plugins.withId("publish-plugin") {
   project.description = "跨平台zip打包模块"
   project.version = "1.2.0"
 }
-
-val isPublish =
-  gradle.startParameter.taskNames.any { it.endsWith("publish") || it.endsWith("publishToMavenLocal") }
 
 kotlin {
   androidTarget {
@@ -128,7 +124,9 @@ uniffi {
 
 tasks.named("compileKotlinDesktop") {
   doFirst {
-    projectDir.resolve("src").deleteRecursively()
+    if (!project.isPublish) {
+      projectDir.resolve("src").deleteRecursively()
+    }
   }
   doLast {
     copyDirectoryToTarget(
@@ -177,7 +175,7 @@ tasks.named("gen-bindings") {
 project.afterEvaluate {
   tasks.named("buildBindings") {
     doLast {
-      if (isPublish) {
+      if (project.isPublish) {
         projectDir.resolve("build").resolve("generated").resolve("uniffi").listFiles().forEach {
           if (it.path.contains("Main")) {
             it.deleteRecursively()
